+++ src/Stackmaster/PatchInstaller.cs
@@ -31,6 +31,10 @@ namespace Stackmaster
                 Postfix(typeof(Player), "HaveRequirementItems", new[] { typeof(Recipe), typeof(bool), typeof(int), typeof(int) }, typeof(NearbyRequirementPatches), "RecipePostfix"),
                 Postfix(typeof(Player), "HaveRequirements", new[] { typeof(Piece), typeof(Player.RequirementMode) }, typeof(NearbyRequirementPatches), "PiecePostfix"),
                 Postfix(typeof(Hud), "SetupPieceInfo", new[] { typeof(Piece) }, typeof(NearbyBuildHudPatch)),
+                Postfix(typeof(InventoryGui), "SetupRequirement", new[]
+                {
+                    typeof(UnityEngine.Transform), typeof(Piece.Requirement), typeof(Player), typeof(bool), typeof(int), typeof(int)
+                }, typeof(NearbyCraftingHudPatch)),
                 Postfix(typeof(Player), "GetFirstRequiredItem", new[]
                 {
                     typeof(Inventory), typeof(Recipe), typeof(int), typeof(int).MakeByRefType(), typeof(int).MakeByRefType(), typeof(int)